Split AddHumansInfrastructure by section to reduce PR merge thrash
The monolithic 378-line InfrastructureServiceCollectionExtensions.cs
registered ~60 services across every section of the app, so every
in-flight PR that added or changed a registration conflicted on the
same file. With ~20 PRs open, the cascading rebases were burning more
time than the features themselves.

Extracts each section's DI registrations into per-section extension
files under Extensions/Sections/ and cross-cutting integrations into
Extensions/Infrastructure/. The public entrypoint
(AddHumansInfrastructure) keeps the same signature; its body is now
a flat list of Add{Section}/Add{Infrastructure} calls.

Zero behavior change: registrations move verbatim (comments, keyed
registrations, Singleton-over-Scoped forwarding, environment-gated
branches). GdprExportDependencyInjectionTests confirm the descriptor
graph is identical (1183/1183 unit tests pass).

Plan: docs/superpowers/plans/2026-04-22-di-split-by-section.md

Follow-up: pr-swarm-rebase the 20 open PRs to move their new
registrations from the orchestrator body into the correct section
file.
