+++ src/Humans.Web/Extensions/Sections/UsersSectionExtensions.cs
@@ -1,0 +1,29 @@
+using Humans.Application.Interfaces;
+using Humans.Application.Interfaces.Gdpr;
+using Humans.Application.Interfaces.Repositories;
+using Humans.Infrastructure.Repositories;
+using Humans.Infrastructure.Services;
+using UsersUserService = Humans.Application.Services.Users.UserService;
+
+namespace Humans.Web.Extensions.Sections;
+
+internal static class UsersSectionExtensions
+{
+    internal static IServiceCollection AddUsersSection(this IServiceCollection services)
+    {
+        // User section — §15 repository pattern (issue #511).
+        // No decorator / cache: User is ~500 rows with no stitched projection or
+        // hot bulk-read path; see docs/superpowers/specs/2026-04-21-issue-511-user-migration.md
+        // for the Option A rationale. IUserRepository is Singleton
+        // (IDbContextFactory-based) so the service can inject it directly.
+        services.AddSingleton<IUserRepository, UserRepository>();
+        services.AddScoped<UsersUserService>();
+        services.AddScoped<IUserService>(sp => sp.GetRequiredService<UsersUserService>());
+        services.AddScoped<IUserDataContributor>(sp => sp.GetRequiredService<UsersUserService>());
+
+        services.AddScoped<IMembershipCalculator, MembershipCalculator>();
+        services.AddScoped<IDashboardService, DashboardService>();
+
+        return services;
+    }
+}
